+++ src/peakrdl_rust/component_context.py
@@ -184,10 +184,13 @@ class ContextScanner(RDLListener):
         byte_endian: Literal["Big", "Little"],
         word_endian: Literal["Big", "Little"],
         access_mode: str = "software",
+        read_only: bool = False,
     ) -> None:
         self.top_nodes = top_nodes
         self.byte_endian: Literal["Big", "Little"] = byte_endian
         self.word_endian: Literal["Big", "Little"] = word_endian
+        self.access_mode = access_mode
+        self.read_only = read_only
         self.top_component_modules: list[str] = []
         self.components: dict[Path, Component] = {}
         self.msg = top_nodes[0].env.msg
@@ -262,7 +265,7 @@ class ContextScanner(RDLListener):
                 addr_offset = child.address_offset
 
             if isinstance(child, RegNode):
-                if not utils.reg_access(child, self.access_mode):
+                if not utils.reg_access(child, self.access_mode, self.read_only):
                     continue
                 registers.append(
                     RegisterInst(
@@ -327,10 +330,10 @@ class ContextScanner(RDLListener):
                 memories=memories,
                 size=node.size,
             )
-        else:
+        else:  # MemNode
             assert len(submaps) == 0
             assert len(memories) == 0
-            if not (access := utils.field_access(node, self.access_mode)):
+            if not (access := utils.mem_access(node, self.access_mode, self.read_only)):
                 return WalkerAction.Continue
             memwidth = node.get_property("memwidth")
             primitive_width = 2 ** int(math.ceil(math.log2(memwidth)))
@@ -431,7 +434,7 @@ class ContextScanner(RDLListener):
                     comment=utils.doc_comment(field),
                     inst_name=snakecase(field.inst_name),
                     type_name=pascalcase(field.inst_name),
-                    access=utils.field_access(field, self.access_mode),
+                    access=utils.field_access(field, self.access_mode, self.read_only),
                     primitive=primitive,
                     encoding=encoding_name,
                     exhaustive=exhaustive,